--- src/main/java/com/example/study/controller/dto/LoginRequest.java
@@ -1,7 +1,0 @@
-package com.example.study.controller.dto;
-
-public record LoginRequest(
-        String loginId,
-        String password
-) {
-}
