--- src/main/java/com/example/study/controller/dto/LoginResponse.java
@@ -1,7 +1,0 @@
-package com.example.study.controller.dto;
-
-public record LoginResponse(
-        String accessToken,
-        String tokenType
-) {
-}
