+++ src/main/java/com/example/study/domain/auth/controller/dto/LoginRequest.java
@@ -1,0 +1,7 @@
+package com.example.study.domain.auth.controller.dto;
+
+public record LoginRequest(
+        String loginId,
+        String password
+) {
+}
