+++ src/main/java/com/example/study/domain/auth/controller/dto/LoginResponse.java
@@ -1,0 +1,7 @@
+package com.example.study.domain.auth.controller.dto;
+
+public record LoginResponse(
+        String accessToken,
+        String tokenType
+) {
+}
