+++ src/main/java/com/example/study/domain/auth/util/AuthorizationUtils.java
@@ -1,0 +1,15 @@
+package com.example.study.domain.auth.util;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class AuthorizationUtils {
+    public static String getAccessToken(String authHeader) {
+        if (authHeader != null && authHeader.startsWith("Bearer ")) {
+            return authHeader.substring(7);
+        }
+        throw new IllegalArgumentException("토큰에 문제가 있습니다.");
+    }
+
+}
