+++ src/main/java/com/example/study/domain/member/controller/dto/MemberRequest.java
@@ -1,0 +1,17 @@
+package com.example.study.domain.member.controller.dto;
+
+import com.example.study.domain.member.entity.MemberEntity;
+
+public record MemberRequest(
+         String loginId,
+         String password,
+         String name
+) {
+    public static MemberEntity toEntity(MemberRequest request) {
+        return new MemberEntity(
+                request.loginId(),
+                request.password(),
+                request.name()
+        );
+    }
+}
