+++ src/main/java/com/example/study/domain/member/controller/dto/MemberResponse.java
@@ -1,0 +1,17 @@
+package com.example.study.domain.member.controller.dto;
+
+import com.example.study.domain.member.entity.MemberEntity;
+
+public record MemberResponse(
+        Long id,
+        String loginId,
+        String name
+) {
+    public static MemberResponse fromEntity(MemberEntity entity) {
+        return new MemberResponse(
+                entity.getId(),
+                entity.getLoginId(),
+                entity.getName()
+        );
+    }
+}
