+++ src/main/java/com/example/study/domain/post/controller/HelloController.java
@@ -1,0 +1,17 @@
+package com.example.study.domain.post.controller;
+
+import com.example.study.domain.post.service.HelloService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class HelloController {
+    private final HelloService helloService;
+
+    @GetMapping("/hello")
+    public String hello() {
+        return helloService.hi();
+    }
+}
