+++ src/main/java/com/example/study/domain/post/controller/dto/PostRequest.java
@@ -1,0 +1,21 @@
+package com.example.study.domain.post.controller.dto;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class PostRequest {
+    private final String title;
+    private final String content;
+    private final String author;
+
+    @Override
+    public String toString() {
+        return "CreatePostRequest{" +
+                "title='" + title + '\'' +
+                ", content='" + content + '\'' +
+                ", author='" + author + '\'' +
+                '}';
+    }
+}
