+++ src/main/java/com/example/study/domain/post/service/HelloService.java
@@ -1,0 +1,11 @@
+package com.example.study.domain.post.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class HelloService {
+
+    public String hi() {
+        return "Hello!";
+    }
+}
