--- src/main/java/com/example/study/entity/MemberEntity.java
@@ -1,25 +1,0 @@
-package com.example.study.entity;
-
-import lombok.Getter;
-
-@Getter
-public class MemberEntity {
-    private static Long AUTO_INCREMENT = 1L;
-
-    private Long id;
-    private String loginId;
-    private String password;
-    private String name;
-
-    public MemberEntity(String loginId, String password, String name) {
-        this.id = AUTO_INCREMENT++;
-        this.loginId = loginId;
-        this.password = password;
-        this.name = name;
-    }
-
-    public boolean isVAlidPassword(String password) {
-        return this.password.equals(password);
-    }
-
-}
