--- src/main/java/com/example/study/service/HelloService.java
@@ -1,11 +1,0 @@
-package com.example.study.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class HelloService {
-
-    public String hi() {
-        return "Hello!";
-    }
-}
